refactor to app delegate to gain reusability
