+++ nef/Extensions/Data+IO.swift
@@ -24,4 +24,14 @@ extension Data {
             return image
         }
     }
+    
+    func persistImage(command: Command, assembler: Assembler = Assembler()) -> IO<AppDelegate.Error, URL> {
+        assembler.resolveOpenPanel().writableFolder(create: true).use { folder in
+            let output = IO<OpenPanelError, URL>.var()
+            return binding(
+                output <- folder.outputURL(command: command),
+                       |<-self.writeIO(to: output.get).mapError { _ in .unknown },
+            yield: output.get)
+        }^.mapError { _ in .carbon }^
+    }
 }
